queryInterface/requirements.txt
streamlit
elasticsearch
pandas
python-dotenv==1.0.1